Move notes to their own resource group
There are enough actions and views to warrant breaking this out of the
application choices controller.
