+++ app/views/provider_interface/notes/show.html.erb
@@ -1,0 +1,10 @@
+<div class="app-notes__note">
+  <h2 class="govuk-heading-m govuk-!-margin-bottom-1">
+    <%= @note.subject %>
+  </h2>
+  <p class="govuk-body govuk-!-margin-top-1 govuk-!-margin-bottom-1"><%= @note.message %></p>
+  <p class="meta">
+    <%= "#{@note.provider_user.full_name}," if @note.provider_user.full_name.present? %>
+    <%= @note.created_at.to_s(:govuk_date_and_time) %>
+  </p>
+</div>
